requirements.txt
numpy>=1.13
tensorflow-gpu>=1.5
tqdm
gym>=0.10
git+https://github.com/Souphis/gym-vrep.git
